fix gamePage
